set storage variables to uint64

+++ contracts/shard/IShardVaultAdmin.sol
@@ -127,7 +127,7 @@ interface IShardVaultAdmin {
      * @notice sets the maxSupply of shards
      * @param maxSupply the maxSupply of shards
      */
-    function setMaxSupply(uint256 maxSupply) external;
+    function setMaxSupply(uint64 maxSupply) external;
 
     /**
      * @notice sets the whitelistEndsAt timestamp
@@ -139,7 +139,7 @@ interface IShardVaultAdmin {
      * @notice sets the maximum amount of shard to be minted during whitelist
      * @param reservedShards reserved shard amount
      */
-    function setReservedShards(uint256 reservedShards) external;
+    function setReservedShards(uint64 reservedShards) external;
 
     /**
      * @notice sets the isEnabled flag, allowing or prohibiting deposits
@@ -162,7 +162,7 @@ interface IShardVaultAdmin {
      * but once this amount is exceeded said user may not deposit more
      * @param maxUserShards new maxUserShards value
      */
-    function setMaxUserShards(uint256 maxUserShards) external;
+    function setMaxUserShards(uint64 maxUserShards) external;
 
     /**
      * @notice unstakes from JPEG'd LPFarming, then from JPEG'd citadel, then from curve LP
